requirements.txt
aiohttp
asyncio
flask
google-generativeai
iewt
nltk
numpy
requests
tflite-runtime
websocket-client